(call_blkid) always bypass blkid cache
This reverts the use of the blkid cache.

This cache is a broken concept and should not be used. It's only
intended to be used for LABEL/UUID conversion.

From the upstream maintainer:
<kzak> coling: -p provides more information, the cache is designed for
LABEL/UUID conversion -- and the goal is to avoid the cache if possible
(it's mostly for backward compatibility). The ideal solution is to read
the information from udev DB.
<kzak> coling: man blkid (at least the latest version contains some hint
about this issue)
<kzak> coling: I'd like to learn people to use lsblk -- it's designed
more friendly  for end-users as well as for scripts and it reads info
from udev, libblkid is only fallback here.

Longer term we should kill off the use of blkid and perhaps move to
lsblk or some perl-udev (if such a thing exists) usage instead:

<kay> coling: avoid the blkid cache, it is a completely broken idea
<kay> kzak: you should really kill that thing :)
<kzak> kay: I'd like to kill blkid at all and keep it as "to test the
library" binary...
<kay> kzak: tools with options like that talk for their sanity
themselves :) " -g     Perform a garbage collection pass on the blkid
cache to remove devices which no longer exist."
<kay> kzak: it's just silly, really silly to ever do that :)
<kay> kzak: yeah, sounds fine to let blkid and its cache die in the long
run
<kzak>  lsblk is maintainable and extendable -- fix blkid(8) is
impossible to fix...

This reverts r6891.

git-svn-id: svn://svn.mageia.org/soft/drakx/trunk@7324 5b7c3c08-40e5-403b-9995-ace06908e4af