Do not use /dev/inside prefix.
As /dev/ is bind mounted into the prefix, we should always refer to /dev/ directly.
In particular this is needed inside wild_device.pm which is often used before the
prefix is even ready (i.e. when configuring existing partitions).

A specific problem here was detecting the major/minor numbers of partitions
which failed because the device node was being searched for inside the prefix
but we'd not even unlocked the encrypted / partition yet to mount it... catch 22.

mga#5661

git-svn-id: svn://svn.mageia.org/soft/drakx/trunk@7908 5b7c3c08-40e5-403b-9995-ace06908e4af