finish-install: Do not check for /usr/sbin/urpmi.update
This file is now 'hidden' in /usr/libexec and only used via it's wrapper
script in mgaonline.

This was originally changed from /usr/bin/urpmi back in 2001 by Pixel, but
the current path for urpmi is /usr/sbin/urpmi, which seems to be
as reliable as urpmi.update was before.

Conflicts:
	perl-install/install/NEWS