fix a crash after "insufficient space" error
create_display_box gets called twice. The first time, only the $kind
parameter has a value, the second time all the parameters (including
$button) have values.

On the first time through the partitioning step, the second call to
create_display_box sets the selected_for_resize flag for my NTFS
partition. When returning to the partitioning step, this flag is still
set, so the first call to create_display_box executes the block of code
that ends with

  $button->signal_connect('focus-in-event' => sub { ... });

even though it doesn't have a value for $button.
(Martin Whitaker, mga#15919)