fix getting systemd services status at install
Some services such as shorewall are not "wantedby".
However, as systemd is not running during installation, we failed to
detect if whether they're enabled or not as the legacy fallback doesn't
handle such services

Let's try hard to query status of services that are not "wanted" but
we're checking for.
Thus fixing updating firewall status after configuration (mga#17288)