fix description of network interfaces (mga#15638)
It got broken when switching stage1 to udev: previously we were manually
loading the driver and thus we knew which driver corresponded to the
newly created intf.

Since mga4, the driver has already been loaded by udev and thus we fail
to associate the intf with the right driver and thus we failed to get
the driver description.

=> let's use the ethtool IOCTL like the stage2 does