Fix erratic behaviour when telling kernel to delete partitions (mga#20074).
When telling the kernel about changes to a DOS partition table, if a
partition was deleted on a disk that also contained an Empty partition,
the kernel also removed the Empty partition from its cached partition
table (and renumbered the other partitions).

Experimentation showed that leaving the udev exec queue active whilst we
were telling the kernel about the changes fixed this problem, although I
don't have an explanation for why it does.