Fix auto-allocation of BIOS boot partitions (mga#20161, mga#19888).
This adds a specific subroutine, fsedit::auto_allocate_boot_bios_parts
that detects if a BIOS boot partition is needed and allocates it if so.
This allows us to relax the rules in fs::any::is_boot_bios_part_needed
to allow the user to manually allocate the BIOS boot partition on a
different device if they so wish.

In the normal case that installation is confined to a single disk,
this will allocate a single BIOS boot partition on that disk. In
the rare case that installation is spread over multiple disks, it
will allocate a BIOS boot partition on every disk. Given that the
BIOS boot partitions are very small and that this is not a normal
use case (see mga#16055), this seems an acceptable quirk - and does
allow the user to then choose any disk when installing the boot
loader.