Allow a BIOS boot partition to be located on any disk (mga#20161).
This relaxes the rules in fs::any::is_boot_bios_part_needed to allow
the BIOS boot partition to be located on any disk, not just the disk
containing the root partition. It also ensures that a BIOS boot
partition is never forcibly required in a system with non-GPT disks.