Don't require a BIOS boot partition when performing an upgrade (mga#20338)
If we are upgrading a system on a GPT disk and it doesn't currently
have a BIOS boot partition, the old system must be using a legacy
bootloader (e.g. grub). We support the continued use of legacy
bootloaders in an upgrade, so there is no need to insist on a BIOS
boot partition in this case.