Fix drakboot video mode selection when using grub2 (mga#21246)
The video mode needs to be extracted from and written back to the
GRUB_CMDLINE_LINUX_DEFAULT parameter in /etc/defaults/grub2.

This fix creates a new "default_vga" key in the $bootloader hash to
hold the current selection. Using the existing "vga" key would cause
the default video mode to be added to any bootloader entries that
don't have a specific "vga" value, e.g. the "failsafe" entry.

The $bootloader "vga" value doesn't appear to be set anywhere else,
so may in fact be obsolete.