installer: force update of /dev/disk/by-uuid after partitioning (mga#22059)
Because stage2 does not include the udev 60-blocks.rule, udev does not
automatically update the soft links in /dev/disk/by-uuid after we write
the partition table and format the partitions. We need these links to
be updated before we create the initrd. It would be cleaner to fix this
with a udev rule, but for now, use brute force.